Remove EncryptableSocket
