+++ src/Socket.php
@@ -17,4 +17,27 @@ interface Socket extends ReadableStream, WritableStream
     public function getLocalAddress(): SocketAddress;
 
     public function getRemoteAddress(): SocketAddress;
+
+    /**
+     * @return void Returns when TLS is successfully set up on the socket.
+     *
+     * @throws SocketException Promise fails and the socket is closed if setting up TLS fails.
+     */
+    public function setupTls(?Cancellation $cancellation = null): void;
+
+    /**
+     * @return void Returns when TLS is successfully shutdown.
+     *
+     * @throws SocketException Promise fails and the socket is closed if shutting down TLS fails.
+     */
+    public function shutdownTls(?Cancellation $cancellation = null): void;
+
+    public function isTlsAvailable(): bool;
+
+    public function getTlsState(): TlsState;
+
+    /**
+     * @return TlsInfo|null The TLS (crypto) context info if TLS is enabled on the socket or null otherwise.
+     */
+    public function getTlsInfo(): ?TlsInfo;
 }
